[Refactor] data 모듈에 있던 DI 관련 클래스 app 모듈로 이전
