+++ app/src/main/java/com/toucheese/di/StudioModule.kt
@@ -15,7 +15,7 @@ import javax.inject.Singleton
 
 @InstallIn(SingletonComponent::class)
 @Module
-object AppModule {
+object StudioModule {
 
 
     @Provides
