--- data/src/main/java/com/tedmoon99/data/di/remote/RetrofitModule.kt
@@ -1,59 +1,0 @@
-package com.tedmoon99.data.di.remote
-
-import com.tedmoon99.data.BuildConfig.BASE_URL
-import com.tedmoon99.data.datasource.remote.studio.api.StudioService
-import dagger.Module
-import dagger.Provides
-import dagger.hilt.InstallIn
-import dagger.hilt.components.SingletonComponent
-import okhttp3.OkHttpClient
-import okhttp3.logging.HttpLoggingInterceptor
-import retrofit2.Retrofit
-import retrofit2.converter.gson.GsonConverterFactory
-import java.util.concurrent.TimeUnit
-import javax.inject.Singleton
-
-@InstallIn(SingletonComponent::class)
-@Module
-object RetrofitModule {
-
-
-    @Provides
-    @Singleton
-    fun provideLoggingInterceptor(): HttpLoggingInterceptor {
-        return HttpLoggingInterceptor().apply {
-            // 요청 및 응답 로그 출력
-            level = HttpLoggingInterceptor.Level.BODY
-        }
-    }
-
-
-    @Provides
-    @Singleton
-    fun provideOkHttpClient(loggingInterceptor: HttpLoggingInterceptor): OkHttpClient {
-        return OkHttpClient
-            .Builder()
-            .connectTimeout(30, TimeUnit.SECONDS) // 서버 연결 대기 시간
-            .readTimeout(30, TimeUnit.SECONDS) // 서버 응답 대기 시간
-            .writeTimeout(30, TimeUnit.SECONDS) // 요청 데이터 전송 시간
-            .addNetworkInterceptor(loggingInterceptor) // 로그 인터셉터
-            .build()
-    }
-
-    @Provides
-    @Singleton
-    fun provideRetrofit(okHttpClient: OkHttpClient): Retrofit {
-        return Retrofit
-            .Builder()
-            .baseUrl(BASE_URL)
-            .client(okHttpClient)
-            .addConverterFactory(GsonConverterFactory.create())
-            .build()
-    }
-
-    @Provides
-    @Singleton
-    fun provideStudioService(retrofit: Retrofit): StudioService {
-        return retrofit.create(StudioService::class.java)
-    }
-}
